discard: Record insertion positions before batching
Discard-to-batch translates selected additions into ownership after loading
either a cached hunk or an explicit file view.

Those paths previously skipped the reference pass, so repeated gaps reached
baseline projection without an original insertion position.

This commit continues the insertion work by recording addition boundaries
on both discard line-selection paths before ownership is prepared.

Subsequent tests will save, reapply, and project those additions across
repeated gaps and shifted indexes.
